Added functionality for hover highlighting code blocks
Requires using an additional argument into the pygments for the code block, this
will wrap each line with an id of `name-[line #]`.  Then any links that target that
anchor id or a range (`name-1:5` targets lines 1-5) will highlight the lines on
hover.  The highlighting is done by adding/removing the `emphasize` class.  There
are still some issues if the code block doesn't have the id-ed spans wrapping each
line, but this could probably be fixed by looking for a `-` in the id of the span
in the CSS (hacky).